requirements.txt
openai>=1.0.0
pandas>=1.3.0
tqdm>=4.65.0
tenacity>=8.2.0
python-dotenv>=0.19.0
pyyaml>=6.0.0
requests>=2.28.0
numpy>=1.20.0
matplotlib>=3.5.0
anthropic>=0.5.0
evals>=0.0.1
typing-extensions>=4.0.0
argparse>=1.4.0
regex>=2022.1.18 